+++ src/main/java/chess/domain/piece/Blank.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.piece;
 
 public class Blank extends Piece {
