+++ src/main/java/chess/domain/piece/Knight.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.piece;
 
 public class Knight extends Piece {
