+++ src/main/java/chess/domain/piece/Queen.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.piece;
 
 public class Queen extends Piece {
